% Part of the oro.dti package for R
% Distributed under the BSD license: see oro.dti/COPYING
%
% $Id: $

\name{imageTensor}
\alias{imageTensor}
\title{Visualizing Fractional Anisotropy and the Principle Eigenvector
  in DTI}
\description{
  ...
}
\usage{
imageTensor(mat3d, weights=NULL, plot=TRUE, add=FALSE, exclude=1, ...)
}
\arguments{
  \item{mat3d}{}
  \item{weights}{}
  \item{plot}{}
  \item{add}{}
  \item{exclude}{}
  \item{...}{}
}
\details{
}
\value{
}
\references{
  Adapted from \code{imagenrgb} by Dr. Agustin Lobo, Instituto de
  Ciencias de la Tierra (CSIC) Lluis Sole Sabaris s/n, 08028 Barcelona
  SPAIN <\email{alobo@ija.csic.es}>
}
%\seealso{}
%\examples{}
\author{Brandon Whitcher <\email{bjw34032@users.sourceforge.net}>}
\keyword{plot}
